Move inspector manager into its own file.
